flex-sensors/requirements.txt
pandas
numpy
pyserial